requirements.txt
python-telegram-bot==22.7
yt-dlp==2026.3.17
APScheduler==3.10.4
